New: Move newznab endpoint to /{id}/api
