+++ src/NzbDrone.Core/Applications/ApplicationBase.cs
@@ -14,7 +14,7 @@ namespace NzbDrone.Core.Applications
         protected readonly IAppIndexerMapService _appIndexerMapService;
         protected readonly Logger _logger;
 
-        protected static readonly Regex AppIndexerRegex = new Regex(@"api\/v\d*\/indexer\/(?<indexer>\d*)",
+        protected static readonly Regex AppIndexerRegex = new Regex(@"(?<indexer>\d*)/api",
                                                                 RegexOptions.IgnoreCase | RegexOptions.Compiled);
 
         public abstract string Name { get; }
